Moved logic to NewtonFractal class
